redirector/app/requirements.txt
Flask==1.1.2
requests==2.23.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
